requirements.txt
eth_typing
safe-eth-py >=7.1.1
